Harden VDEX and DEX structural analysis
Why
- Malformed DEX tables, verifier payloads, and string metadata could trigger excessive work, ambiguous parsing, or unsafe output behavior.
- DEX 041 containers and legacy VDEX variants need explicit format boundaries instead of best-effort assumptions.

What
- Validate DEX file, container, map, fixed-table, reference, checksum, signature, and MUTF-8 invariants while preserving recovery diagnostics.
- Bound verifier, type-lookup, string, and explain-field work; handle valid isolated UTF-16 surrogates without 32-bit allocation overflow.
- Make extraction collision-safe with atomic file creation, sanitize terminal output, and reject incompatible modify versions.

Direction
- Keep outer VDEX analysis recovery-oriented while rejecting unsafe or structurally ambiguous embedded data.
- Report unsupported CompactDex and content-integrity state explicitly rather than implying full support.

Validation
- Go 1.25.11 full tests, vet, race tests, 386 compile, and linux/darwin/windows cross-builds passed.
- Coverage gate passed at 85.3%; golangci-lint v2.12.2 reported 0 issues.
- Android 16 corpus tests and a 643260-byte framework VDEX analysis passed.

+++ pkg/vdex/api_test.go
@@ -156,7 +156,7 @@ func buildMultiDexVDEX(t *testing.T, count int) []byte {
 	pu32(buf, 52, uint32(len(buf)))
 	pu32(buf, 56, 0)
 	for i := 0; i < count; i++ {
-		pu32(buf, headerSize+i*4, uint32(0xDEAD0000+i))
+		pu32(buf, headerSize+i*4, uint32(0xDEAD0000)+uint32(i))
 		copy(buf[dexOffset+i*len(dex):], dex)
 	}
 	return buf
